requirements.txt
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
pillow